Add display to show the sun's position
Displays the sun's position in a quarter-circle arc on the
bottom right corner of the watch.
The sun moves from a 9'oclock to 12'oclock position and back
to show a full 24 hours. The 9'oclock position represents the
solar midnight and 12'oclock position represents the solar noon.
A line marks the daily sun horizon, or the time of both the
sun rise and sun set.
At a glance, you can see how much of the day consists of sunlight,
and where the sun currently is in this time frame.